test: UserInputCarMoveCountFactory test code 작성
1. 너무 큰 숫자 입력 시 예외
2. 숫자 아닌 값 입력 시 예외
3. 숫자에 맞는 객체 생성

+++ src/main/java/dto/UserInputCarMoveCountDto.java
@@ -2,13 +2,13 @@ package dto;
 
 public class UserInputCarMoveCountDto {
 
-    private int moveCount;
+    private long moveCount;
 
-    public UserInputCarMoveCountDto(int moveCount){
+    public UserInputCarMoveCountDto(long moveCount){
         this.moveCount = moveCount;
     }
 
-    public int getMoveCount(){
+    public long getMoveCount(){
         return moveCount;
     }
 }
